# docs/swagger.yaml
basePath: /api
definitions:
  handlers.AuthRequest:
    description: Структура для входа пользователя
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword
        type: string
    type: object
...
  handlers.MerchInfo:
    properties:
      price:
        type: integer
      type:
        type: string
    type: object
  handlers.SendMoney:
    properties:
      coin:
        type: integer
      toUser:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для магазина с авторизацией, покупками и админ-панелью
  title: Shop API
  version: "1.0"
paths:
  /api/admin/merch/new:
    post:
      consumes:
      - application/json
      description: Позволяет добавить новый мерч или изменить цену существующего мерча.
        Проверяет корректность данных и наличие мерча.
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.MerchInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Мерч успешно добавлен или цена обновлена
          schema:
            type: string
        "400":
          description: Некорректное тело запроса, неверный тип или цена мерча
          schema:
            type: string
        "404":
          description: Мерч с таким именем уже существует
          schema:
            type: string
        "500":
          description: Ошибка добавления нового мерча или обновления цены
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Добавление или изменение цены мерча
      tags:
      - Admin
  /api/admin/users/{username}:
    post:
      consumes:
      - application/json
      description: Позволяет перевести монеты работнику по его никнейму, проверяя
        корректность данных и существование получателя.
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SendMoney'
      produces:
      - application/json
      responses:
        "200":
          description: Перевод монет успешен
          schema:
            type: string
        "400":
          description: Некорректное тело запроса или неверное количество монет
          schema:
            type: string
        "404":
          description: Не найден работник или кошелек получателя
          schema:
            type: string
        "500":
          description: Ошибка обновления баланса получателя или фиксации транзакции
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Перевод монет работнику
      tags:
      - Admin
  /api/auth:
    post:
      consumes:
      - application/json
      description: Авторизует пользователя по email и паролю, создавая учетную запись
        автоматически, если пользователя нет в базе.
      parameters:
      - description: Тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает JWT-токен
          schema:
            additionalProperties:
              type: string
            type: object
        "201":
          description: Пользователь создан автоматически
          schema:
            type: string
        "400":
          description: Некорректное тело запроса
          schema:
            type: string
        "401":
          description: Неверный пароль
          schema:
            type: string
        "408":
          description: Запрос отменен клиентом
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Авторизация пользователя
      tags:
      - Auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Отзывает текущий токен авторизации и добавляет его в список отозванных
        токенов.
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Выход выполнен успешно
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Отсутствует токен авторизации / Некорректный формат токена
          schema:
            type: string
        "500":
          description: Не удалось отозвать токен
          schema:
            type: string
      summary: Выход из системы
      tags:
      - Auth
...
swagger: "2.0"